+++ internal/tofu/context_plan.go
@@ -98,9 +98,9 @@ type PlanOpts struct {
 	// will be added to the plan graph.
 	ImportTargets []*ImportTarget
 
-	// EndpointsToRemove are the list of resources and modules to forget from
+	// RemoveStatements are the list of resources and modules to forget from
 	// the state.
-	EndpointsToRemove []addrs.ConfigRemovable
+	RemoveStatements []*refactoring.RemoveStatement
 
 	// GenerateConfig tells OpenTofu where to write any generated configuration
 	// for any ImportTargets that do not have configuration already.
@@ -196,6 +196,10 @@ func (c *Context) Plan(ctx context.Context, config *configs.Config, prevRunState
 	// includes language asking the user to report a bug.
 	varDiags := checkInputVariables(config.Module.Variables, opts.SetVariables)
 	diags = diags.Append(varDiags)
+
+	// Already having all the variables' values figured out, we can now warn on the user if it's using
+	// variables that are deprecated
+	diags = diags.Append(warnOnUsedDeprecatedVars(opts.SetVariables, config.Module.Variables))
 
 	if len(opts.Targets) > 0 || len(opts.Excludes) > 0 {
 		diags = diags.Append(tfdiags.Sourceless(
@@ -331,7 +335,7 @@ func (c *Context) plan(ctx context.Context, config *configs.Config, prevRunState
 	}
 
 	var endpointsToRemoveDiags tfdiags.Diagnostics
-	opts.EndpointsToRemove, endpointsToRemoveDiags = refactoring.GetEndpointsToRemove(config)
+	opts.RemoveStatements, endpointsToRemoveDiags = refactoring.FindRemoveStatements(config)
 	diags = diags.Append(endpointsToRemoveDiags)
 
 	if diags.HasErrors() {
@@ -856,7 +860,7 @@ func (c *Context) planGraph(config *configs.Config, prevRunState *states.State, 
 			ExternalReferences:      opts.ExternalReferences,
 			ImportTargets:           opts.ImportTargets,
 			GenerateConfigPath:      opts.GenerateConfigPath,
-			EndpointsToRemove:       opts.EndpointsToRemove,
+			RemoveStatements:        opts.RemoveStatements,
 			ProviderFunctionTracker: providerFunctionTracker,
 		}).Build(addrs.RootModuleInstance)
 		return graph, walkPlan, diags
@@ -1130,4 +1134,27 @@ func (c *Context) relevantResourceAttrsForPlan(config *configs.Config, plan *pla
 	}
 
 	return contributors, diags
+}
+
+// warnOnUsedDeprecatedVars is checking for variables whose values are given by the user and if any of that is
+// marked as deprecated, a warning message is written for it.
+func warnOnUsedDeprecatedVars(inputs InputValues, decls map[string]*configs.Variable) (diags tfdiags.Diagnostics) {
+	for vn, in := range inputs {
+		if in.SourceType == ValueFromConfig {
+			continue
+		}
+		vc, ok := decls[vn]
+		if !ok {
+			continue
+		}
+		if vc.Deprecated != "" {
+			diags = diags.Append(&hcl.Diagnostic{
+				Severity: hcl.DiagWarning,
+				Summary:  "Deprecated variable used from the root module",
+				Detail:   fmt.Sprintf("The root variable %q is deprecated with the following message: %s", vc.Name, vc.Deprecated),
+				Subject:  vc.DeclRange.Ptr(),
+			})
+		}
+	}
+	return diags
 }